Mooshak g5 60%

+++ pegajales/g1.c
@@ -27,10 +27,15 @@
  */
 struct elemento operador(struct elemento x, struct elemento y, char *op)
 {
-    float dx, dy, dres;
+    double dx, dy, dres;
+
+    //printf("Antes do convertTOdouble temos x : %d e y: %d\n", x.data.val_i, y.data.val_i);
 
     dx = convertToDouble(x);
     dy = convertToDouble(y);
+
+    // printf("Na funcção operador temos o dx a %f e o dy a %f\n", dx, dy);
+
     dres = 0.0;
     switch (*op)
     {
@@ -73,7 +78,7 @@ struct elemento operador(struct elemento x, struct elemento y, char *op)
  * 
  * @returns Um struct elemento val
  */
-struct elemento secondoperador(struct elemento x, struct elemento y, float dres)
+struct elemento secondoperador(struct elemento x, struct elemento y, double dres)
 {
     struct elemento val;
