+++ pegajales/g1.h
@@ -36,7 +36,7 @@ struct elemento operador(struct elemento x, struct elemento y, char *op);
  * 
  * @returns Um struct elemento val
  */
-struct elemento secondoperador(struct elemento x, struct elemento y, float dres);
+struct elemento secondoperador(struct elemento x, struct elemento y, double dres);
 
 /**
  * \brief Função convertToDouble do programa
@@ -47,9 +47,9 @@ struct elemento secondoperador(struct elemento x, struct elemento y, float dres)
  * 
  * @param xs
  * 
- * @returns Um float
+ * @returns Um double
  */
-float convertToDouble(struct elemento x);
+long double convertToDouble(struct elemento x);
 
 /**
  * \brief Função secondoperador
